separate GraphNodeEditor into own class
